Remove obsolete shared service indirection for #142
The client-side API client is no longer present, so the shared service factory layer only adds indirection around server-only generated API usage. This folds the affected feature services into their server entrypoints and keeps generated response data unwrapping explicit at the service boundary.

Server Action error handling preserves backend messages best-effort without defining a new frontend-owned error contract, because the generated API output does not expose a typed failure DTO contract.

Constraint: Issue #142 scope includes service consolidation, unwrapData, and conservative action error handling
Constraint: Generated API output does not expose a typed failure DTO contract
Rejected: Strong frontend error DTO standardization | backend failure shape is not guaranteed by generated contracts
Rejected: Regenerate API client | not needed for this refactor and outside clarified scope
Confidence: high
Scope-risk: moderate
Directive: Do not broaden frontend error-code mapping without a generated or backend-confirmed failure contract
Tested: npm run lint; npx tsc --noEmit; JEST_SKIP_MSW=true CI=true npm test -- --runInBand --forceExit; npm run build
Not-tested: npm test -- --runInBand with MSW enabled hangs after starting Jest in this workspace; pre-commit lint-staged hangs before spawning eslint/prettier in this environment
Related: #142
